# [GenevaTraceExporter] Use the stack to avoid needing to enumerate tags twice