## Changes

* Uses the stack to avoid enumerating tags twice will exporting traces.

## Benchmarks

(The perf depends on the number of tags. For this benchmark I used 6.)

Before:

|            Method |     Mean |   Error |  StdDev | Allocated |
|------------------ |---------:|--------:|--------:|----------:|
|    ExportActivity | 417.2 ns | 5.04 ns | 4.71 ns |         - |
| SerializeActivity | 377.6 ns | 3.26 ns | 2.89 ns |         - |

After:

|            Method |     Mean |   Error |  StdDev | Allocated |
|------------------ |---------:|--------:|--------:|----------:|
|    ExportActivity | 342.6 ns | 2.53 ns | 2.24 ns |         - |
| SerializeActivity | 323.1 ns | 4.11 ns | 3.64 ns |         - |

## TODOs

* [ ] Appropriate `CHANGELOG.md` updated for non-trivial changes