@CodeBlanch 2022-12-19
Opened as a draft because there are a couple unknowns I need to chase down:

* This will reverse the order of tags written to `env_properties`, is that safe to do?
* How many tags can we handle before the stack is exhausted?

Wanted to see if anyone else had feedback while I work on that.