@utpilla 2022-12-23
> * How many tags can we handle before the stack is exhausted?

How about having another `byte[]` for serializing `env_properties` as we enumerate over tags. After the end of the enumeration over the tags, we would just have to copy the contents of this `byte[]` to the original buffer. This approach would be more reliable. In the worst case, we would get a catchable `IndexOutOfRangeException` if there are too many `env_properties` to serialize. That would still be better than hitting a potential `StackOverflowException` and getting the process crashed.

We could definitely modify the recursion method to keep a threshold for when to stop recursing further but even from an implementation and maintenance standpoint, having an additional buffer and copying the contents over to the original one seems much simpler.